Add files via upload
I put these three texts into a single folder and. Only the last same part of the file name is important, since the first function take advantage of this same part to go through every name of files in this folder and then plus the same main link (HTTP://.......), then implemented them in the for loop and to do the argument classification.